Removed date_required_by and reason_for_priority, added priority_level to sample message
