+++ tests/data/examples_create_labware_messages.py
@@ -35,8 +35,6 @@ TEST_CREATE_LABWARE_MSG_OBJECT: Dict[str, Any] = {
                 "shearingAndQCComments": "Comments",
                 "dateSubmittedUTC": datetime.now(),
                 "priorityLevel": "Medium",
-                "dateRequiredBy": "Long Read",
-                "reasonForPriority": "Reason goes here",
             },
             {
                 "sampleUuid": "dd490ee5-fd1d-456d-99fd-eb9d3861e0f7".encode(),
@@ -65,8 +63,6 @@ TEST_CREATE_LABWARE_MSG_OBJECT: Dict[str, Any] = {
                 "shearingAndQCComments": "Comments",
                 "dateSubmittedUTC": datetime.now(),
                 "priorityLevel": "Medium",
-                "dateRequiredBy": "Long Read",
-                "reasonForPriority": "Reason goes here",
             },
         ],
     },
@@ -107,8 +103,6 @@ TEST_INVALID_CREATE_LABWARE_MSG_OBJECT: Dict[str, Any] = {
                 "shearingAndQCComments": "Comments",
                 "dateSubmittedUTC": datetime.now(),
                 "priorityLevel": "Medium",
-                "dateRequiredBy": "Long Read",
-                "reasonForPriority": "Reason goes here",
             },
             {
                 "sampleUuid": "dd490ee5-fd1d-456d-99fd-eb9d3861e0f7".encode(),
@@ -137,8 +131,6 @@ TEST_INVALID_CREATE_LABWARE_MSG_OBJECT: Dict[str, Any] = {
                 "shearingAndQCComments": "Comments",
                 "dateSubmittedUTC": datetime.now(),
                 "priorityLevel": "Medium",
-                "dateRequiredBy": "Long Read",
-                "reasonForPriority": "Reason goes here",
             },
         ],
     },
