+++ tol_lab_share/constants/input_create_labware_message.py
@@ -35,5 +35,4 @@ FINAL_NANODROP = "finalNanoDrop"
 SHEARING_QC_COMMENTS = "shearingAndQCComments"
 DATE_SUBMITTED_UTC = "dateSubmittedUTC"
 PRIORITY_LEVEL = "priorityLevel"
-DATE_REQUIRED_BY = "dateRequiredBy"
-REASON_FOR_PRIORITY = "reasonForPriority"
+
